backend/requirements.txt
django
djangorestframework
psycopg2
django-rest-framework-social-oauth2 